+++ src/main/java/by/itacademy/jd2/service/impl/PositionServiceImpl.java
@@ -15,6 +15,9 @@ import by.itacademy.jd2.service.api.PositionService;
 import by.itacademy.jd2.utils.PaginatorUtil;
 
 import jakarta.transaction.Transactional;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import java.io.Serializable;
 import java.sql.Date;
 import java.time.LocalDate;
@@ -23,6 +26,7 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 public class PositionServiceImpl implements PositionService {
+    private static final Logger LOGGER = LoggerFactory.getLogger(PositionServiceImpl.class);
     private static final boolean DEFAULT_IS_ACTUAL = true;
     private final PositionDAO positionDAO;
     private final DepartmentDAO departmentDAO;
@@ -32,7 +36,6 @@ public class PositionServiceImpl implements PositionService {
         this.departmentDAO = new DepartmentDaoImpl();
     }
 
-    @Transactional
     @Override
     public void addPosition(PositionDTO positionDTO) {
         PositionEntity positionEntity = Converter.toEntity(positionDTO, PositionEntity.class);
@@ -40,9 +43,10 @@ public class PositionServiceImpl implements PositionService {
         positionEntity.setDepartment(departmentEntity);
         departmentEntity.getPositions().add(positionEntity);
         positionDAO.save(positionEntity);
+
+        LOGGER.info("Added position with id={}", positionEntity.getId());
     }
 
-    @Transactional
     @Override
     public void updatePosition(PositionDTO positionDTO) {
         if (positionDTO != null) {
@@ -51,31 +55,45 @@ public class PositionServiceImpl implements PositionService {
             newPosition.setDepartment(oldPosition.getDepartment());
             newPosition.setIsActual(oldPosition.getIsActual());
             positionDAO.update(newPosition, newPosition.getId());
+
+            LOGGER.info("Updated position with id={}", newPosition.getId());
         }
     }
 
-    @Transactional
     @Override
     public void deletePosition(Serializable id) {
         PositionEntity positionEntity = positionDAO.get(id);
-        positionEntity.getDepartment().getPositions().remove(positionEntity);
-        positionDAO.delete(id);
+        if (positionEntity != null) {
+            positionEntity.getDepartment().getPositions().remove(positionEntity);
+            positionDAO.delete(id);
+
+            LOGGER.info("Deleted position with id={}", id);
+        }
     }
 
     @Override
     public PositionDTO getPosition(Serializable id) {
-        return Converter.toDto(positionDAO.get(id), PositionDTO.class);
+        PositionDTO positionDTO = Converter.toDto(positionDAO.get(id), PositionDTO.class);
+
+        if (positionDTO != null) {
+            LOGGER.info("Found position with id={}", id);
+        }
+
+        return positionDTO;
     }
 
     @Override
     public List<PositionItemDTO> getAllPositionItems() {
-        return positionDAO.getAll().stream()
+        List<PositionItemDTO> items = positionDAO.getAll().stream()
                 .filter(PositionEntity::getIsActual)
                 .map(entity -> Converter.toDto(entity, PositionItemDTO.class))
                 .collect(Collectors.toList());
+
+        LOGGER.info("Found {} position items", items.size());
+
+        return items;
     }
 
-    @Transactional
     @Override
     public PageInfo<PositionDTO> getPositionsByDepartmentIdAndActualAndPage(Serializable departmentId,
                                                                             Boolean isActual,
@@ -97,10 +115,11 @@ public class PositionServiceImpl implements PositionService {
                 .orElse(null);
         Long positionsCount = positionDAO.getPositionsCountByDepartmentIdAndActual(departmentId, isActual);
 
+        LOGGER.info("Found {} from {} positions", positions.size(), positionsCount);
+
         return new PageInfo<>(positions, pageNumber, pageSize, positionsCount);
     }
 
-    @Transactional
     @Override
     public void reducePosition(Serializable id) {
         PositionEntity position = positionDAO.get(id);
@@ -113,11 +132,15 @@ public class PositionServiceImpl implements PositionService {
                 });
         positionDAO.update(position, position.getId());
         this.changePositionStatus(id, false);
+
+        LOGGER.info("Reduced position with id={}", id);
     }
 
     @Override
     public void restorePosition(Serializable id) {
         this.changePositionStatus(id, true);
+
+        LOGGER.info("Restored position with id={}", id);
     }
 
     @Override
